examples/spark-connect-unit-tests-docker-demo/requirements.txt
pyspark-connect==4.0.1
pytest==8.3.4
pytest-timeout==2.3.1
psycopg2-binary==2.9.9
